WIP: SQL Server support
* refactor customerror (WIP)
* more code in query but still non-functional
